# additup/WebApi/WebApi/Model/RankEnum.cs
/* 
 * AddItUp
 *
 * Add It Up API Definitions
 *
 * OpenAPI spec version: 1.0.0
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Text.RegularExpressions;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.Runtime.Serialization;
using Newtonsoft.Json;
using Newtonsoft.Json.Converters;
using System.ComponentModel.DataAnnotations;

namespace WebApi.Model
{
    /// <summary>
    /// Possible Ranks
    /// </summary>
    /// <value>Possible Ranks</value>
    [JsonConverter(typeof(StringEnumConverter))]
    public enum RankEnum
    {
        
        /// <summary>
        /// Enum Beginner for "Beginner"
        /// </summary>
        [EnumMember(Value = "Beginner")]
        Beginner,
        
        /// <summary>
        /// Enum Talented for "Talented"
        /// </summary>
        [EnumMember(Value = "Talented")]
        Talented,
        
        /// <summary>
        /// Enum Intermediate for "Intermediate"
        /// </summary>
        [EnumMember(Value = "Intermediate")]
        Intermediate,
        
        /// <summary>
        /// Enum Advanced for "Advanced"
        /// </summary>
        [EnumMember(Value = "Advanced")]
        Advanced,
        
        /// <summary>
        /// Enum Expert for "Expert"
        /// </summary>
        [EnumMember(Value = "Expert")]
        Expert
    }

}