# additup/WebApiClient/additup-app/additup-ui/src/app/datamodels/Answer.ts
/**
 * AddItUp
 * Add It Up API Definitions
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import * as models from './models';

export class Answer {
    sessionId: string;

    exerciseId: string;

    submittedAnswer: string;

}